6.move classes to main packages.
